platform

+++ NativeLib/CommonLib.cpp
@@ -1,6 +1,6 @@
 #include "CommonLib.h"
 
-int clamp(int x, int min, int max)
+float clamp(float x, float min, float max)
 {
 	return (x < min) ? min : ((x > max) ? max : x);
 }
